@@ -1,49 +1,28 @@
 package pointer
 
-func ToString(str string) *string {
-	return &str
+func To[T any](t T) *T {
+	return &t
 }
 
-func ToInt(num int) *int {
-	return &num
+func ToOrNil[T comparable](t T) *T {
+	if z, ok := interface{}(t).(interface{ IsZero() bool }); ok {
+		if z.IsZero() {
+			return nil
+		}
+		return &t
+	}
+
+	var zero T
+	if t == zero {
+		return nil
+	}
+	return &t
 }
 
-func ToInt8(num int8) *int8 {
-	return &num
-}
-
-func ToInt16(num int16) *int16 {
-	return &num
-}
-
-func ToInt32(num int32) *int32 {
-	return &num
-}
-
-func ToInt64(num int64) *int64 {
-	return &num
-}
-
-func ToUint(num uint) *uint {
-	return &num
-}
-
-func ToUint8(num uint8) *uint8 {
-	return &num
-}
-
-func ToUint16(num uint16) *uint16 {
-	return &num
-}
-
-func ToUint32(num uint32) *uint32 {
-	return &num
-}
-
-func ToUint64(num uint64) *uint64 {
-	return &num
-}
-
-func ToBool(val bool) *bool {
-	return &val
+func Get[T interface{}](t *T) T {
+	if t == nil {
+		var zero T
+		return zero
+	}
+	return *t
 }
